Create a chain object

Create an object that describes mooring chain elements such as shackles, either by looking up a known object from the database, or by defining a new type. Area is formulated as length*width, for consistency with Dewey's Matlab code.

chain(
  model = "1in buoy chain",
  buoyancyPerMeter = NULL,
  areaPerMeter = NULL,
  CD = NULL,
  length = NULL
)

Arguments

model

character value indicating the type of chain. There are three possibilities for model. (1) If this is "?", then the function returns a vector of permitted character values. (2) If this is "?X", where X is a set of characters, then findElement() is called to do a fuzzy search, with the search argument set to "chain". (3) If it is recognized, i.e. if that type is stored in data(mooringElements), then the that stored value is used, and all other arguments to this function are ignored. (4) Otherwise, a new chain object is created, using values specified in the other arguments (all of which must be supplied).

buoyancyPerMeter

numeric value indicating buoyancy per meter of length, expressed as kg/m. Note that buoyancyPerMeter is ignored if model is recognized.

areaPerMeter

numeric value indicating the frontal area per meter of length, expressed as m^2. (This is equivalent to the width, or diameter, of the wire.) Note that areaPerMeter is ignored if model is recognized.

CD

numeric value (unitless) for the drag coefficient. Note that CD is ignored if model is recognized.

length

(mandatory) numeric value indicating the length (in m) of the wire.

Value

chain returns an object of the "mooring" class and "chain" subclass.

Data sources

To learn more about the data used for the built-in chain objects, use e.g.

data(mooringElements)
mooringElements$chains

See also

Author

Dan Kelley

Examples

library(mooring)
# List known chain types
chain("?")
#>  [1] "1 chain SL"             "1/2in buoy chain"       "1/2in chain LL"        
#>  [4] "1/2in chain SL"         "1/4in chain SL"         "1in buoy chain"        
#>  [7] "3/4in buoy chain"       "3/8in buoy chain"       "3/8in chain SL"        
#> [10] "5/8in buoy chain"       "5/8in galvanized chain"