Clamp an Instrument on a Wire

This changes the height field of an instrument object to 0, so that it does not contribute to the overall length of the mooring. Using this in deep water is unlikely to make much difference, but it is handy for shallow-water cases, if the device was clamped on to the wire, but the height field of the instrument object used to represent it has non-zero height. The example illustrates this, for mooring in very shallow water.

clamped(instrument)

Arguments

instrument

a mooring object of subtype instrument, as created with instrument().

Value

clamped returns a copy of its input, but with the height value set to 0.

Author

Dan Kelley

Examples

library(mooring)
a <- anchor("my anchor", buoyancy = -50, height = 0.1, CD = 0)
w <- function(length) wire(length = length)
i <- instrument("SBE37 microcat clamp-on style")
f <- float("my float", buoyancy = 20, height = 0.2, area = 0.2^2, CD = 1.3)
# Construct unclamped mooring (m) and clamped mooring (M).
mooring(a, w(20), i, w(20), f, waterDepth = 50)
#> <mooring::mooringS7>
#>  @ elements  :List of 5
#>  .. $ : <mooring::floatS7>
#>  ..  ..@ model       : chr "my float"
#>  ..  ..@ buoyancy    : num 20
#>  ..  ..@ height      : num 0.2
#>  ..  ..@ area        : num 0.04
#>  ..  ..@ CD          : num 1.3
#>  ..  ..@ source      : chr ""
#>  ..  ..@ originalName: chr ""
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -9.14
#>  ..  ..@ z0          : num -9.14
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 20
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::wireS7>
#>  ..  ..@ model       : chr "1/4in wire/jack"
#>  ..  ..@ buoyancy    : num -2.6
#>  ..  ..@ height      : num 20
#>  ..  ..@ area        : num 0.16
#>  ..  ..@ CD          : num 1.3
#>  ..  ..@ source      : chr "Dewey"
#>  ..  ..@ originalName: chr "1/4 wire/jack"
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -9.34
#>  ..  ..@ z0          : num -9.34
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 17.4
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::instrumentS7>
#>  ..  ..@ model       : chr "SBE37 microcat clamp-on style"
#>  ..  ..@ buoyancy    : num -1.53
#>  ..  ..@ height      : num 0.559
#>  ..  ..@ area        : num 0.117
#>  ..  ..@ CD          : num 0.65
#>  ..  ..@ source      : chr "BIO"
#>  ..  ..@ originalName: chr "SBE37 MICROCAT CLAMP-ON STYLE"
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -29.3
#>  ..  ..@ z0          : num -29.3
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 15.9
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::wireS7>
#>  ..  ..@ model       : chr "1/4in wire/jack"
#>  ..  ..@ buoyancy    : num -2.6
#>  ..  ..@ height      : num 20
#>  ..  ..@ area        : num 0.16
#>  ..  ..@ CD          : num 1.3
#>  ..  ..@ source      : chr "Dewey"
#>  ..  ..@ originalName: chr "1/4 wire/jack"
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -29.9
#>  ..  ..@ z0          : num -29.9
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 13.3
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::anchorS7>
#>  ..  ..@ model       : chr "my anchor"
#>  ..  ..@ buoyancy    : num -50
#>  ..  ..@ height      : num 0.1
#>  ..  ..@ area        : num 0
#>  ..  ..@ CD          : num 0
#>  ..  ..@ source      : chr ""
#>  ..  ..@ originalName: chr ""
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -49.9
#>  ..  ..@ z0          : num -49.9
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 13.3
#>  ..  ..@ group       : num 0
#>  @ waterDepth: num 50
#>  @ u         : num 0
mooring(a, w(20), clamped(i), w(20), f, waterDepth = 50)
#> <mooring::mooringS7>
#>  @ elements  :List of 5
#>  .. $ : <mooring::floatS7>
#>  ..  ..@ model       : chr "my float"
#>  ..  ..@ buoyancy    : num 20
#>  ..  ..@ height      : num 0.2
#>  ..  ..@ area        : num 0.04
#>  ..  ..@ CD          : num 1.3
#>  ..  ..@ source      : chr ""
#>  ..  ..@ originalName: chr ""
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -9.7
#>  ..  ..@ z0          : num -9.7
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 20
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::wireS7>
#>  ..  ..@ model       : chr "1/4in wire/jack"
#>  ..  ..@ buoyancy    : num -2.6
#>  ..  ..@ height      : num 20
#>  ..  ..@ area        : num 0.16
#>  ..  ..@ CD          : num 1.3
#>  ..  ..@ source      : chr "Dewey"
#>  ..  ..@ originalName: chr "1/4 wire/jack"
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -9.9
#>  ..  ..@ z0          : num -9.9
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 17.4
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::instrumentS7>
#>  ..  ..@ model       : chr "SBE37 microcat clamp-on style"
#>  ..  ..@ buoyancy    : num -1.53
#>  ..  ..@ height      : num 0
#>  ..  ..@ area        : num 0.117
#>  ..  ..@ CD          : num 0.65
#>  ..  ..@ source      : chr "BIO"
#>  ..  ..@ originalName: chr "SBE37 MICROCAT CLAMP-ON STYLE"
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -29.9
#>  ..  ..@ z0          : num -29.9
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 15.9
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::wireS7>
#>  ..  ..@ model       : chr "1/4in wire/jack"
#>  ..  ..@ buoyancy    : num -2.6
#>  ..  ..@ height      : num 20
#>  ..  ..@ area        : num 0.16
#>  ..  ..@ CD          : num 1.3
#>  ..  ..@ source      : chr "Dewey"
#>  ..  ..@ originalName: chr "1/4 wire/jack"
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -29.9
#>  ..  ..@ z0          : num -29.9
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 13.3
#>  ..  ..@ group       : num 0
#>  .. $ : <mooring::anchorS7>
#>  ..  ..@ model       : chr "my anchor"
#>  ..  ..@ buoyancy    : num -50
#>  ..  ..@ height      : num 0.1
#>  ..  ..@ area        : num 0
#>  ..  ..@ CD          : num 0
#>  ..  ..@ source      : chr ""
#>  ..  ..@ originalName: chr ""
#>  ..  ..@ x           : num 0
#>  ..  ..@ x0          : num 0
#>  ..  ..@ z           : num -49.9
#>  ..  ..@ z0          : num -49.9
#>  ..  ..@ phi         : num 0
#>  ..  ..@ tau         : num 13.3
#>  ..  ..@ group       : num 0
#>  @ waterDepth: num 50
#>  @ u         : num 0