Drag on mooring elements

This looks up element areas with area() and drag coefficients with CD(), then computes drag force (in Newtons) with \((1/2)*area*rho*CD*u^2\) and divides by g to get a mass equivalance, which is returned.

drag(m, u, rho = 1027, g = 9.8)

Arguments

m

either a mooring object, created by mooring(), or a mooring-element object, created with

u

either a non-negative number indicating depth-independent velocity, or a function that returns that velocity as a function of depth (m) below the ocean surface.

rho

a numeric value giving water density in kg/m^3. The default value of 1027 is likely to be sufficient for most calculations, given that oceanographic variations of density are typically well under 1 percent, although it might make sense to use a smaller value in lake applications. Note that the overall force balance is more sensitive to other things, as discussed in the “Details” section of the documentation for buoyancy().

g

numeric value of the acceleration due to gravity, with default being 9.8 m/s^2.

Value

drag returns a numeric vector of horizontal drag "force" (really, force divided by gravitational acceleration), expressed in kg.

Details

Although fluid density rho and g are parameters to this function, the default values are likely to be used in all practical oceanographic calculations, because drag coefficient is not known to three digits.

Author

Dan Kelley