Create a float object

Create a float object. Some notes follow.

HMB in a name is a short-hand for Hydrofloat Mooring Buoy. Data for these floats was extracted from Reference 1 on 2021-05-19 by Dan Kelley, with the CD value being set at 0.65 (in the absence of any data from the manufacture) because that value is used in Dewey's dataset for many floats.

float(
  model = "Kiel SFS40in",
  buoyancy = NULL,
  height = NULL,
  diameter = NULL,
  CD = NULL
)

Arguments

model
character value indicating the model name of the float. The default is "Kiel SFS40in". Use float("?") to get a listing of permitted values. If model is not recognized, then float creates a new type, using the stated diameter and CD values; otherwise, these arguments are ignored.

buoyancy
numeric value of the buoyancy of the float, in kg. This is ignored if model is recognized.

height
numeric value for the height of the element, in m. This is ignored if model is recognized.

diameter
numeric value for the diameter of the float, in m. This is ignored if model is recognized.

CD
numeric value for the drag coefficient for the item.

Value

float returns an object of the "mooring" class, with type equal to "float".

References

https://deepwaterbuoyancy.com/wp-content/uploads/hydro-float-mooring-buoys-deepwater-buoyancy.pdf

See also

Author

Dan Kelley

Examples

library(mooring)
# List known float types
float("?")

#>  [1] "16in Viny"        "17 in glass"      "28in ORE"         "30in float"      
#>  [5] "37in ORE"         "41in ORE"         "48in ORE"         "54cm Alum CAP3"  
#>  [9] "54x30inClamShell" "61in ORE"         "ADCP FloatTech"   "BENTHOS 17in"    
#> [13] "Billings-12in"    "BPS Float"        "CDMS 1.2m sphere" "CDMS 2m sphere"  
#> [17] "Double 16inViny"  "Double 17in"      "FloatTech CF-12"  "Geodyn 8' Toroid"
#> [21] "HMB 20"           "HMB 25"           "HMB 31"           "HMB 36"          
#> [25] "HMB 40"           "HMB 44"           "HMB 49"           "HMB 51"          
#> [29] "HMB 56"           "HMB 62"           "Kiel SFS40in"     "Top of Tow Rope" 
#> [33] "Torp. flt, 3-28"  "Tripple FT CF-12" "Trpl 12in glass"  "Trpl 16 in Viny"