Create an instrument object

Create an instrument object, either by looking up a known object from the database, or by defining a new type.

instrument(
  model = "SBE37 microcat clamp-on style",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)

Arguments

model

character value indicating the type of instrument. There are three possibilities for model. (1) If this is "?", then the function returns a vector of permitted character values. (2) If this is "?X", where X is a set of characters, then findElement() is called to do a fuzzy search, with the search argument set to "instrument". (3) If it is recognized, i.e. if that type is stored in data(mooringElements), then the that stored value is used, and all other arguments to this function are ignored. (4) Otherwise, a new instrument object is created, using values specified in the other arguments (all of which must be supplied).

buoyancy

numeric value indicating buoyancy. For wire elements, this is buoyancy per length, expressed in kg/m. For other elements, it is buoyancy, expressed in kg. Note that buoyancy is ignored if model is recognized.

height

numeric value for the height (in m) of the element. This value is used in the computation of the mooring geometry. In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated as zero, it is also used in the computation of the area of the element, as height*width. Note that height is ignored if model is recognized.

area

a numerical value indicating the horizontal instrument area in \(m^2\).

CD

numeric value (unitless) for the drag coefficient. Note that CD is ignored if model is recognized.

Value

instrument returns an object of the "mooring" class and "instrument" subclass.

Data sources

To learn more about the data used for the built-in instrument objects, use e.g.

data(mooringElements)
mooringElements$instruments

See also

Author

Dan Kelley

Examples

library(mooring)
# List known instrument types
instrument("?")
#>  [1] "ADCP sentinel (orange) 500m with inline frame" 
#>  [2] "ADCP sentinel (yellow) 6000m with inline frame"
#>  [3] "Aanderaa RCM-4"                                
#>  [4] "Aanderaa RCM-5"                                
#>  [5] "Aanderaa RCM-7"                                
#>  [6] "Aanderaa RCM-8"                                
#>  [7] "Aanderaa RDCP"                                 
#>  [8] "CDMS CM"                                       
#>  [9] "DVS current meter"                             
#> [10] "EG&G VACM"                                     
#> [11] "EG&G VMCM"                                     
#> [12] "FSI-3DACM coast"                               
#> [13] "Hobo DO U26"                                   
#> [14] "Hobo Temp U22"                                 
#> [15] "InterOcn S4 CM"                                
#> [16] "NBIS ACM-2"                                    
#> [17] "OEI Expend. CM"                                
#> [18] "RCM-11 in frame"                               
#> [19] "RCM-8 with fin"                                
#> [20] "RD ADCP"                                       
#> [21] "RDI with bar"                                  
#> [22] "SBE37 microcat clamp-on style"                 
#> [23] "SBE37 microcat on a ss bar"                    
#> [24] "VR2AR reciever"                                
#> [25] "WH ADCP + Frame"                               
#> [26] "aquaMeasure CHLA"                              
#> [27] "aquaMeasure DOT"                               
#> [28] "aquaMeasure SAL"                               
#> [29] "seabird CTD (ios oxygen with bar)"             
#> [30] "seacat 16-03 (bar,plastic case)"               
#> [31] "seacat 16-04 (bar,titanium no press.)"         
#> [32] "seacat 16-04 (bar,titanium with press.)"       
#> [33] "seaguard (2000 & 6000 m)"                      
#> [34] "sediment trap"                                 
#> [35] "vemco on a ss bar"                             
#> [36] "wotan (bar)"