Create a misc object

Create a miscellaneous object, either by looking up a known object from the database, or by defining a new type. The function name, and most of the built-in data values, come from Dewey's (1999, 2021) database. For data derived from the Dewey database, area is computed as in Dewey's Matlab code: if tabulated diameter is zero, then the product of tabulated height and width is used; otherwise, pi*(diameter/2)^2 is used.

misc(
  model = "AanderaaT.chain",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL,
  source = NULL,
  originalName = originalName
)

Arguments

model

character value indicating the type of misc. There are three possibilities for model. (1) If this is "?", then the function returns a vector of permitted character values. (2) If this is "?X", where X is a set of characters, then findElement() is called to do a fuzzy search, with the search argument set to "misc". (3) If it is recognized, i.e. if that type is stored in data(mooringElements), then the that stored value is used, and all other arguments to this function are ignored. (4) Otherwise, a new misc object is created, using values specified in the other arguments (all of which must be supplied).

buoyancy

numeric value indicating buoyancy. For wire elements, this is buoyancy per length, expressed in kg/m. For other elements, it is buoyancy, expressed in kg. Note that buoyancy is ignored if model is recognized.

height

numeric value for the height (in m) of the element. This value is used in the computation of the mooring geometry. In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated as zero, it is also used in the computation of the area of the element, as height*width. Note that height is ignored if model is recognized.

area

a numerical value indicating the horizontal instrument area in \(m^2\).

CD

numeric value (unitless) for the drag coefficient. Note that CD is ignored if model is recognized.

source

character value indicating the source of the data.

originalName

character value indicating the original model name.

Value

misc returns an object of the "mooringElement" class and "misc" subclass.

See also

Author

Dan Kelley

Examples

library(mooring)
# List known misc types
misc("?")
#>  [1] "A. tide gauge"    "AanderaaT.chain"  "BPS Package"      "Benthos Xpdr"    
#>  [5] "Brancker TR1000"  "Brancker XL-200"  "Brancker XL-210"  "CDMS 10sqft Tow" 
#>  [9] "CDMS 30sqft Tow"  "DataSonics Xpdr"  "Heavy Sphere"     "Heavy Sphere 2"  
#> [13] "IceProfiler4+Flt" "Keil sedmt trap"  "McLane Mk78 SdTr" "McLane Mk8 SedTr"
#> [17] "ONC-uCatADop"     "Oceanetic Pinger" "SBE 19+ cage"     "SBE-16"          
#> [21] "SubOcean2d"       "SubOcean2d2"      "TowedCylinder"    "TowedSphere"     
#> [25] "ULS#2/APL"