Get Mooring/Element side-view Area

area returns a numeric vector of mooring element side-view area(s), in square metres, taking into account the orientation of the element with respect to the vertical. (Floats are assumed to be spherical, so the orientation is irrelevant. This is not the case for wires and instruments.) The return value will be a single number if m is a single element but in a typical mooring it will have multiple values. The areas are listed with the top element first, i.e. in the reverse order to that used in constructing the mooring with mooring().

area(m, phi = TRUE)

Arguments

m

either a mooring object, created by mooring(), or a mooring-element object, created with anchor(), chain(), connector(), instrument(), misc(), release(), or wire().

phi

a numeric or logical value that controls how the angle is computed. If this is FALSE, no angle is used. If it is TRUE, which is the default, then the value of phi within the mooring element is used (or 0 is used, if there is no value). Finally, if phi is a single numerical value, then that is taken as the angle to the vertical, in radians.

Value

area returns a numeric value of the side-view area, in m^2.

Details

When initially created, objects do not contain an angle value, but such values are inserted by knockdown(). Moorings that are significantly tilted will have reduced side-view area, which will reduce the current-induced drag, as computed with drag().

Author

Dan Kelley

Examples

library(mooring)
# Floats are unaffected by rotation
area(float())
#> [1] 0.7853982
area(float(), pi / 4)
#> [1] 0.7853982
# Other elements, e.g. instruments, are affected by rotation
area(instrument())
#> [1] 0.117
area(instrument(), pi / 4.0) * sqrt(2.0)
#> [1] 0.117