Create a float object

Create a float object

float(
  model = "Kiel SFS40in",
  buoyancy = NULL,
  height = NULL,
  area = NULL,
  CD = NULL
)

Arguments

model

character value indicating the type of float There are three possibilities for model. (1) If this is "?", then the function returns a vector of permitted character values. (2) If this is "?X", where X is a set of characters, then findElement() is called to do a fuzzy search, with the search argument set to "float (3) If it is recognized, i.e. if that type is stored in data(mooringElements)`, then the that stored value is used, and all other arguments to this function are ignored. (4) Otherwise, a new float values specified in the other arguments (all of which must be supplied).

buoyancy

numeric value indicating buoyancy. For wire elements, this is buoyancy per length, expressed in kg/m. For other elements, it is buoyancy, expressed in kg. Note that buoyancy is ignored if model is recognized.

height

numeric value for the height (in m) of the element. This value is used in the computation of the mooring geometry. In some cases, as e.g. in Dewey-derived data that have "diameter" tabulated as zero, it is also used in the computation of the area of the element, as height*width. Note that height is ignored if model is recognized.

area

a numerical value indicating the horizontal instrument area in \(m^2\).

CD

numeric value (unitless) for the drag coefficient. Note that CD is ignored if model is recognized.

Value

float returns a "mooringElement" object with "float" subclass.

References

https://deepwaterbuoyancy.com/wp-content/uploads/hydro-float-mooring-buoys-deepwater-buoyancy.pdf

See also

Author

Dan Kelley

Examples

library(mooring)
# List known float types
float("?")
#>  [1] "11in centre hole tfloat"             
#>  [2] "14in centre hole tfloat"             
#>  [3] "16in Viny"                           
#>  [4] "17in glass"                          
#>  [5] "28in ORE"                            
#>  [6] "3 pack Viny 12b-3 floats"            
#>  [7] "30in float"                          
#>  [8] "37in ORE"                            
#>  [9] "41in ORE"                            
#> [10] "48in ORE"                            
#> [11] "54cm Alum CAP3"                      
#> [12] "54x30inClamShell"                    
#> [13] "61in ORE"                            
#> [14] "8in centre hole tfloat"              
#> [15] "ADCP / 2x c3 subs assembly"          
#> [16] "ADCP FloatTech"                      
#> [17] "BENTHOS 17in"                        
#> [18] "BPS Float"                           
#> [19] "BUB 2x17in glass"                    
#> [20] "Billings 12in"                       
#> [21] "CDMS 1.2m sphere"                    
#> [22] "CDMS 2m sphere"                      
#> [23] "Double 16in Viny"                    
#> [24] "Double 17in"                         
#> [25] "FloatTech CF-12"                     
#> [26] "Geodyn 8ft Toroid"                   
#> [27] "HMB 20"                              
#> [28] "HMB 25"                              
#> [29] "HMB 31"                              
#> [30] "HMB 36"                              
#> [31] "HMB 40"                              
#> [32] "HMB 44"                              
#> [33] "HMB 49"                              
#> [34] "HMB 51"                              
#> [35] "HMB 56"                              
#> [36] "HMB 62"                              
#> [37] "IPS / 2x b3 subs assembly"           
#> [38] "Kiel SFS40in"                        
#> [39] "Top of Tow Rope"                     
#> [40] "Torp. flt, 3-28"                     
#> [41] "Tripple FT CF-12"                    
#> [42] "Trpl 12in glass"                     
#> [43] "Trpl 16in Viny"                      
#> [44] "a2 package ADCP and 2 Viny balls"    
#> [45] "new glass streamlined float c2"      
#> [46] "stablemoor 1000lb 3500 msw with ADCP"
#> [47] "stablemoor 1015lb 1500 msw with ADCP"
#> [48] "stablemoor 533lb 3500 msw with ADCP" 
#> [49] "stablemoor 580lb 1500 msw with ADCP" 
#> [50] "streamlined BUB 2x 17in glass"       
#> [51] "streamlined BUB 3 Viny balls"        
#> [52] "syn. float with ADCP bracket"        
#> [53] "syn. float, bracket and 109lb ADCP"