Create a float object

Create a float object, either by looking up a known object from the database, or defining a new type.

float(
  model = "Kiel SFS40in",
  buoyancy = NULL,
  height = NULL,
  diameter = NULL,
  CD = NULL
)

Arguments

model
character value indicating the type of float. There are three possibilities for model. (1) If this is "?", then the function returns a vector of permitted character values. (2) If it is recognized, i.e. if that type is stored in data(mooringElements), then the that stored value is used, and all other arguments to this function are ignored. (3) Otherwise, a new float object is created, using values specified in the other arguments (all of which must be supplied).

buoyancy
numeric value indicating buoyancy. For wire elements, this is buoyancy per length, expressed in kg/m. For other elements, it is buoyancy, expressed in kg. Note that buoyancy is ignored if model is recognized.

height
numeric value for the height (in m) of the element. This value is used in the computation of (a) the mooring geometry and (b) the area of the element (which is in turn used in the drag computation). Note that height is ignored if model is recognized.

diameter
numeric value for the diameter (in m) of the element as viewed from a horizontal direction. This is used to compute area, for drag calculations, as pi*(diameter/2)^2. Note that diameter is ignored if model is recognized.

CD
numeric value (unitless) for the drag coefficient. Note that CD is ignored if model is recognized.

Value

float returns an object of the "mooring" class and "float" subclass.

Details

Note that HMB in a name is a short-hand for Hydrofloat Mooring Buoy. Data for these floats was extracted from Reference 1 on 2021-05-19 by Dan Kelley, with the CD value being set at 0.65 (in the absence of any data from the manufacture) because that value is used in Dewey's dataset for many floats.

References

https://deepwaterbuoyancy.com/wp-content/uploads/hydro-float-mooring-buoys-deepwater-buoyancy.pdf

See also

Author

Dan Kelley

Examples

library(mooring)
# List known float types
float("?")

#>  [1] "16in Viny"                           
#>  [2] "17in glass"                          
#>  [3] "28in ORE"                            
#>  [4] "3 pack Viny 12b-3 floats"            
#>  [5] "30in float"                          
#>  [6] "37in ORE"                            
#>  [7] "41in ORE"                            
#>  [8] "48in ORE"                            
#>  [9] "54cm Alum CAP3"                      
#> [10] "54x30inClamShell"                    
#> [11] "61in ORE"                            
#> [12] "a2 package adcp and 2 Viny balls"    
#> [13] "ADCP / 2x c3 subs assembly"          
#> [14] "ADCP FloatTech"                      
#> [15] "BENTHOS 17in"                        
#> [16] "Billings 12in"                       
#> [17] "BPS Float"                           
#> [18] "bub 2x17in glass"                    
#> [19] "CDMS 1.2m sphere"                    
#> [20] "CDMS 2m sphere"                      
#> [21] "Double 16in Viny"                    
#> [22] "Double 17in"                         
#> [23] "FloatTech CF-12"                     
#> [24] "Geodyn 8ft Toroid"                   
#> [25] "HMB 20"                              
#> [26] "HMB 25"                              
#> [27] "HMB 31"                              
#> [28] "HMB 36"                              
#> [29] "HMB 40"                              
#> [30] "HMB 44"                              
#> [31] "HMB 49"                              
#> [32] "HMB 51"                              
#> [33] "HMB 56"                              
#> [34] "HMB 62"                              
#> [35] "IPS / 2x b3 subs assembly"           
#> [36] "Kiel SFS40in"                        
#> [37] "stablemoor 1000lb 3500 msw with ADCP"
#> [38] "stablemoor 1015lb 1500 msw with ADCP"
#> [39] "stablemoor 533lb 3500 msw with ADCP" 
#> [40] "stablemoor 580lb 1500 msw with ADCP" 
#> [41] "streamlined bub 2 x 17in glass"      
#> [42] "streamlined bub 3 Viny balls"        
#> [43] "syn. float,bracket and 109lb ADCP"   
#> [44] "syntactic float with ADCP bracket"   
#> [45] "Top of Tow Rope"                     
#> [46] "Torp. flt, 3-28"                     
#> [47] "Tripple FT CF-12"                    
#> [48] "Trpl 12in glass"                     
#> [49] "Trpl 16in Viny"