Dataset of wire, chain, and float properties

This dataset was constructed from the mdcodes.mat (Dewey 1999, 2021). This is provided for use by wire(), chain() and float(). Users will seldom need to consult this table directly, but those that do should note that Dewey's lengths in cm have been converted to m.

data(mooringElements)

References

Dewey, Richard K. "Mooring Design & Dynamics-a Matlab® Package for Designing and Analyzing Oceanographic Moorings." Marine Models 1, no. 1 (December 1, 1999): 103–57. https://doi.org/10.1016/S1369-9350(00)00002-X Dewey, Richard. "Mooring Design and Dynamics." Accessed May 15, 2021. http://canuck.seos.uvic.ca/rkd/mooring/moordyn.php

Examples

library(mooring)
data(mooringElements)
mooringElements

#> $anchors
#>                 name buoyancy height code source      originalName
#> 1    1 Railway Wheel     -350   0.15    1  Dewey   1 Railway Wheel
#> 2   2 Railway Wheels     -700   0.35    1  Dewey  2 Railway Wheels
#> 3   3 Railway Wheels    -1050   0.55    1  Dewey  3 Railway Wheels
#> 4   4 Railway Wheels    -1400   0.70    1  Dewey  4 Railway Wheels
#> 5   1 Concrete Block     -200   0.50    1  Dewey  1 Concrete Block
#> 6  2 Concrete Blocks     -400   0.70    1  Dewey 2 Concrete Blocks
#> 7  3 Concrete Blocks     -600   1.00    1  Dewey 3 Concrete Blocks
#> 8      1 Tire-St+Con     -115   0.15    1  Dewey     1 Tire-St+Con
#> 9  1 Large St/C Tire     -150   0.15    1  Dewey 1 Large St/C Tire
#> 10     3 trainwheels    -1050   0.00   NA kelley              <NA>
#> 
#> $chains
#>                      name buoyancyPerMeter  width  CD code source
#> 1  5/8in galvanized chain           -4.323 0.0572 1.3   NA    BIO
#> 2          1/4in chain SL           -1.690 0.0150 1.3    1  Dewey
#> 3          3/8in chain SL           -2.330 0.0220 1.3    1  Dewey
#> 4          1/2in chain SL           -4.120 0.0300 1.3    1  Dewey
#> 5              1 chain SL          -13.000 0.0400 1.3    1  Dewey
#> 6          1/2in chain LL           -3.100 0.0300 1.3    1  Dewey
#> 7        3/8in buoy chain           -1.677 0.0200 1.3    1  Dewey
#> 8        1/2in buoy chain           -3.070 0.0250 1.3    1  Dewey
#> 9        5/8in buoy chain           -4.730 0.0350 1.3    1  Dewey
#> 10       3/4in buoy chain           -6.870 0.0400 1.3    1  Dewey
#> 11         1in buoy chain          -12.250 0.0500 1.3    1  Dewey
#>              originalName
#> 1  5/8"  GALVANIZED CHAIN
#> 2            1/4 chain SL
#> 3            3/8 chain SL
#> 4            1/2 chain SL
#> 5              1 chain SL
#> 6            1/2 chain LL
#> 7         3/8" buoy chain
#> 8         1/2" buoy chain
#> 9         5/8" buoy chain
#> 10        3/4" buoy chain
#> 11          1" buoy chain
#> 
#> $connectors
#>                        name buoyancy height  width  CD code source
#> 1                    swivel   -0.953  0.150 0.0300 1.3   NA    BIO
#> 2            7/16in shackle   -0.150  0.000 0.0000 1.3   NA    BIO
#> 3    1/4in nicopress sleeve   -0.034  0.000 0.0000 1.3   NA    BIO
#> 4  5/16in heavy SS thimbles   -0.041  0.000 0.0000 1.3   NA    BIO
#> 5            ballast weight  -45.408  1.000 0.0254 1.3   NA    BIO
#> 6            shac-ring-shac   -0.750  0.300 0.0250 1.3    1  Dewey
#> 7           shac-3link-shac   -1.000  0.400 0.0200 1.3    1  Dewey
#> 8             1/2in SS link   -0.200  0.100 0.0250 1.3    1  Dewey
#> 9           1/2in galv link   -0.200  0.100 0.0250 1.3    1  Dewey
#> 10            1/2in shackle   -0.300  0.080 0.0250 1.3    1  Dewey
#> 11            3/8in shackle   -0.125  0.040 0.0250 1.3    1  Dewey
#> 12            5/8in shackle   -0.650  0.065 0.0500 1.3    1  Dewey
#> 13              1in shackle   -2.200  0.080 0.0300 1.3    1  Dewey
#> 14             1/2in swivel   -0.650  0.140 0.0500 1.3    1  Dewey
#> 15             3/4in swivel   -1.500  0.140 0.0400 1.3    1  Dewey
#> 16    1/2in shac+3/8in shac   -0.425  0.120 0.0250 1.3    1  Dewey
#> 17           3T SeineSwivel   -0.650  0.105 0.0400 1.3    1  Dewey
#> 18            3/4in shackle   -0.880  0.060 0.0250 1.3    1  Dewey
#> 19          shac-Pring-shac   -0.500  0.200 0.0200 1.3    1  Dewey
#>               originalName
#> 1                   SWIVEL
#> 2            7/16" SHACKLE
#> 3    1/4" NICOPRESS SLEEVE
#> 4  5/16" HEAVY SS THIMBLES
#> 5           BALLAST WEIGHT
#> 6           shac-ring-shac
#> 7          shac-3link-shac
#> 8              1/2 SS link
#> 9            1/2 galv link
#> 10             1/2 shackle
#> 11             3/8 shackle
#> 12             5/8 shackle
#> 13              1" shackle
#> 14              1/2 swivel
#> 15              3/4 swivel
#> 16        1/2 shac+3/8shac
#> 17          3T SeineSwivel
#> 18            3/4" shackle
#> 19         shac-Pring-shac
#> 
#> $floats
#>                                    name buoyancy height diameter   CD
#> 1        new glass streamlined float c2    45.41 0.8700   0.4820 0.60
#> 2         streamlined BUB 2x 17in glass    45.41 0.7400   0.4550 0.60
#> 3      a2 package ADCP and 2 Viny balls    29.08 0.5630   0.3970 0.60
#> 4              3 pack Viny 12b-3 floats    57.14 1.0000   0.5290 1.10
#> 5          streamlined BUB 3 Viny balls    52.65 0.5630   0.3970 0.60
#> 6                      BUB 2x17in glass    43.78 1.1870   0.8390 1.10
#> 7             IPS / 2x b3 subs assembly    78.88 1.0000   0.7480 0.60
#> 8            ADCP / 2x c3 subs assembly    42.24 1.0000   0.9100 0.60
#> 9   stablemoor 533lb 3500 msw with ADCP   192.45 1.2860   1.7270 0.60
#> 10 stablemoor 1000lb 3500 msw with ADCP   404.39 1.2860   1.7270 0.60
#> 11 stablemoor 1015lb 1500 msw with ADCP   411.22 1.2860   1.7200 0.60
#> 12  stablemoor 580lb 1500 msw with ADCP   213.78 1.2860   1.7270 0.60
#> 13         syn. float with ADCP bracket   839.18 2.4200   1.2540 0.60
#> 14   syn. float, bracket and 109lb ADCP   789.80 2.4200   1.2540 0.60
#> 15                             61in ORE   999.00 1.6500   1.6500 0.65
#> 16                             48in ORE   620.00 1.2340   1.2340 0.65
#> 17                             41in ORE   433.00 1.0400   1.0400 0.65
#> 18                             37in ORE   300.00 0.9400   0.9400 0.65
#> 19                             28in ORE   133.00 0.7100   0.7100 0.65
#> 20                           30in float   160.00 0.7600   0.7600 0.65
#> 21                         Kiel SFS40in   320.00 1.0000   1.0000 0.65
#> 22                      Torp. flt, 3-28   364.00 0.8300   0.8300 0.75
#> 23                           17in glass    23.00 0.5700   0.5700 0.95
#> 24                        Billings 12in    10.00 0.3430   0.3430 0.95
#> 25                         BENTHOS 17in    23.00 0.5100   0.5100 0.95
#> 26                            16in Viny    19.00 0.4000   0.4000 0.60
#> 27                          Double 17in    48.00 0.5000   0.6500 0.95
#> 28                     Double 16in Viny    38.00 0.5000   0.6000 0.60
#> 29                       Trpl 16in Viny    56.00 1.5000   0.7500 0.60
#> 30                      Trpl 12in glass    30.00 0.9500   0.6200 1.00
#> 31                     CDMS 1.2m sphere   682.00 1.2000   1.2000 0.50
#> 32                       CDMS 2m sphere  2470.00 2.0000   2.0000 0.50
#> 33                       54cm Alum CAP3    61.00 0.5500   0.5400 0.65
#> 34                     54x30inClamShell   405.00 1.4000   0.7500 0.15
#> 35                      Top of Tow Rope     0.01 0.0000   0.0000 0.65
#> 36                            BPS Float  8000.00 1.5000   0.0000 1.60
#> 37                       ADCP FloatTech   460.00 1.4200   1.4200 0.65
#> 38                      FloatTech CF-12    11.00 0.3000   0.0000 1.00
#> 39                     Tripple FT CF-12    33.00 0.9000   0.0000 1.00
#> 40                    Geodyn 8ft Toroid  1950.00 0.7620   0.0000 0.65
#> 41                               HMB 20    34.50 0.5080   0.5080 0.65
#> 42                               HMB 25    68.50 0.6350   0.6350 0.65
#> 43                               HMB 31   141.00 0.7874   0.7874 0.65
#> 44                               HMB 36   225.00 0.9144   0.9144 0.65
#> 45                               HMB 40   300.00 1.0160   1.0160 0.65
#> 46                               HMB 44   413.00 1.1176   1.1176 0.65
#> 47                               HMB 49   565.00 1.2446   1.2446 0.65
#> 48                               HMB 51   633.00 1.2954   1.2954 0.65
#> 49                               HMB 56   831.00 1.4224   1.4224 0.65
#> 50                               HMB 62  1122.00 1.5748   1.5748 0.65
#>                                code source
#> 1  NEW GLASS STREAMLINED FLOAT (C2)       
#> 2                              <NA>    BIO
#> 3                              <NA>    BIO
#> 4                              <NA>    BIO
#> 5                              <NA>    BIO
#> 6                              <NA>    BIO
#> 7                              <NA>    BIO
#> 8                              <NA>    BIO
#> 9                              <NA>    BIO
#> 10                             <NA>    BIO
#> 11                             <NA>    BIO
#> 12                             <NA>    BIO
#> 13                             <NA>    BIO
#> 14                             <NA>    BIO
#> 15                                1  Dewey
#> 16                                1  Dewey
#> 17                                1  Dewey
#> 18                                1  Dewey
#> 19                                1  Dewey
#> 20                                1  Dewey
#> 21                                1  Dewey
#> 22                                1  Dewey
#> 23                                1  Dewey
#> 24                                1  Dewey
#> 25                                1  Dewey
#> 26                                1  Dewey
#> 27                                1  Dewey
#> 28                                1  Dewey
#> 29                                1  Dewey
#> 30                                1  Dewey
#> 31                                1  Dewey
#> 32                                1  Dewey
#> 33                                1  Dewey
#> 34                                1  Dewey
#> 35                                1  Dewey
#> 36                                1  Dewey
#> 37                                1  Dewey
#> 38                                1  Dewey
#> 39                                1  Dewey
#> 40                                3  Dewey
#> 41                             <NA> Kelley
#> 42                             <NA> Kelley
#> 43                             <NA> Kelley
#> 44                             <NA> Kelley
#> 45                             <NA> Kelley
#> 46                             <NA> Kelley
#> 47                             <NA> Kelley
#> 48                             <NA> Kelley
#> 49                             <NA> Kelley
#> 50                             <NA> Kelley
#>                             originalName
#> 1                                       
#> 2        STREAMLINED BUB  2 x 17"  GLASS
#> 3       A2 PACKAGE ADCP AND 2 VINY BALLS
#> 4               3 pack VINY 12B-3 floats
#> 5          STREAMLINED BUB  3 VINY BALLS
#> 6                      BUB  2x17"  GLASS
#> 7              IPS / 2x B3 SUBS ASSEMBLY
#> 8             ADCP / 2x C3 SUBS ASSEMBLY
#> 9   Stablemoor 533 lb 3500 msw with ADCP
#> 10 Stablemoor 1000 lb 3500 msw with ADCP
#> 11 Stablemoor 1015 lb 1500 msw with ADCP
#> 12  Stablemoor 580 lb 1500 msw with ADCP
#> 13     SYNTACTIC FLOAT WITH ADCP BRACKET
#> 14   SYN. FLOAT, BRACKET AND 109 LB.ADCP
#> 15                              61in ORE
#> 16                              48in ORE
#> 17                              41in ORE
#> 18                              37in ORE
#> 19                              28in ORE
#> 20                            30in float
#> 21                          Kiel SFS40in
#> 22                       Torp. flt, 3-28
#> 23                           17 in glass
#> 24                         Billings-12in
#> 25                          BENTHOS 17in
#> 26                             16in Viny
#> 27                           Double 17in
#> 28                       Double 16inViny
#> 29                       Trpl 16 in Viny
#> 30                       Trpl 12in glass
#> 31                      CDMS 1.2m sphere
#> 32                        CDMS 2m sphere
#> 33                        54cm Alum CAP3
#> 34                      54x30inClamShell
#> 35                       Top of Tow Rope
#> 36                             BPS Float
#> 37                        ADCP FloatTech
#> 38                       FloatTech CF-12
#> 39                      Tripple FT CF-12
#> 40                      Geodyn 8' Toroid
#> 41                                  <NA>
#> 42                                  <NA>
#> 43                                  <NA>
#> 44                                  <NA>
#> 45                                  <NA>
#> 46                                  <NA>
#> 47                                  <NA>
#> 48                                  <NA>
#> 49                                  <NA>
#> 50                                  <NA>
#> 
#> $instruments
#>                                              name buoyancy height    area   CD
#> 1                 seacat 16-03 (bar,plastic case)    -7.49 0.8400 0.12200 0.65
#> 2           seacat 16-04 (bar,titanium no press.)   -10.89 0.8763 0.06200 0.65
#> 3         seacat 16-04 (bar,titanium with press.)   -13.39 1.0668 0.07700 0.65
#> 4               seabird CTD (ios oxygen with bar)   -18.16 1.2700 0.09400 0.65
#> 5                               DVS current meter    -4.54 0.7000 0.24300 0.65
#> 6                                     wotan (bar)   -20.42 1.6800 0.17600 0.65
#> 7                                    RDI with bar   -45.39 2.1800 0.27000 0.65
#> 8                                  RCM-8 with fin   -21.33 0.7400 0.06000 0.65
#> 9                                 RCM-11 in frame   -20.31 0.8000 0.07800 0.65
#> 10                       seaguard (2000 & 6000 m)   -19.59 0.8600 0.08300 0.65
#> 11                              vemco on a ss bar    -2.22 1.2000 0.05200 0.65
#> 12                     SBE37 microcat on a ss bar    -4.65 1.2000 0.06000 0.65
#> 13                                  sediment trap   -61.27 1.8290 1.06100 0.65
#> 14                  SBE37 microcat clamp-on style    -1.53 0.5590 0.11700 0.65
#> 15  ADCP sentinel (orange) 500m with inline frame   -14.69 0.7600 0.09100 0.65
#> 16 ADCP sentinel (yellow) 6000m with inline frame   -22.04 0.7900 0.11100 0.65
#> 17                                 Aanderaa RCM-4   -17.30 0.5500 0.07040 1.30
#> 18                                 Aanderaa RCM-5   -19.50 0.5700 0.07296 1.30
#> 19                                 Aanderaa RCM-7   -18.30 0.5500 0.07040 1.30
#> 20                                 Aanderaa RCM-8   -22.70 0.5700 0.07296 1.30
#> 21                                 OEI Expend. CM    -3.60 0.6600 0.08448 1.30
#> 22                                FSI-3DACM coast   -10.00 0.3800 0.04940 1.30
#> 23                                      EG&G VACM   -35.00 1.9800 0.49500 1.30
#> 24                                      EG&G VMCM   -30.00 1.7000 0.42500 1.30
#> 25                                     NBIS ACM-2   -13.00 1.2000 0.24000 1.30
#> 26                                 InterOcn S4 CM    -1.50 0.8000 0.00000 0.60
#> 27                                        RD ADCP   -12.00 1.4100 0.31020 1.30
#> 28                                        CDMS CM   -20.00 2.0000 0.40000 1.20
#> 29                                WH ADCP + Frame   -12.00 0.5000 0.10000 1.20
#> 30                                  Aanderaa RDCP   -22.00 0.5800 0.09280 1.30
#>    code source                                          originalName
#> 1    NA    BIO                      SEACAT 16-03 (BAR, PLASTIC CASE)
#> 2    NA    BIO                 SEACAT 16-04 (BAR,TITANIUM NO PRESS.)
#> 3    NA    BIO               SEACAT 16-04 (BAR,TITANIUM WITH PRESS.)
#> 4    NA    BIO                     SEABIRD CTD (IOS OXYGEN WITH BAR)
#> 5    NA    BIO                                     DVS CURRENT METER
#> 6    NA    BIO                                           WOTAN (BAR)
#> 7    NA    BIO                                          RDI WITH BAR
#> 8    NA    BIO                                        RCM-8 WITH FIN
#> 9    NA    BIO                                       RCM-11 IN FRAME
#> 10   NA    BIO                              SEAGUARD (2000 & 6000 M)
#> 11   NA    BIO                                     VEMCO ON A SS BAR
#> 12   NA    BIO                            SBE37 MICROCAT ON A SS BAR
#> 13   NA    BIO                                 SEDIMENT TRAP (#1349)
#> 14   NA    BIO                         SBE37 MICROCAT CLAMP-ON STYLE
#> 15   NA    BIO  WHADCP Sentinel (orange) 500 meter with inline frame
#> 16   NA    BIO WHADCP Sentinel (yellow) 6000 meter with inline frame
#> 17    1  Dewey                                        Aanderaa RCM-4
#> 18    1  Dewey                                        Aanderaa RCM-5
#> 19    1  Dewey                                        Aanderaa RCM-7
#> 20    1  Dewey                                        Aanderaa RCM-8
#> 21    1  Dewey                                        OEI Expend. CM
#> 22    1  Dewey                                       FSI-3DACM coast
#> 23    1  Dewey                                             EG&G VACM
#> 24    1  Dewey                                             EG&G VMCM
#> 25    1  Dewey                                            NBIS ACM-2
#> 26    1  Dewey                                        InterOcn S4 CM
#> 27    1  Dewey                                               RD ADCP
#> 28    1  Dewey                                               CDMS CM
#> 29    3  Dewey                                       WH ADCP + Frame
#> 30    1  Dewey                                         Aanderaa RDCP
#> 
#> $miscs
#>                 name buoyancy height   CD code       area source
#> 1  AanderaaT.chain    -12.000  0.550 1.30    1 0.07040000  Dewey
#> 2  SBE-16             -10.000  0.450 1.30    1 0.03600000  Dewey
#> 3  A. tide gauge      -12.000  0.550 1.30    1 0.07040000  Dewey
#> 4  Keil sedmt trap    -80.000  2.500 1.30    1 2.49750000  Dewey
#> 5  ULS#2/APL           15.000  0.000 1.00    1 0.33183072  Dewey
#> 6  DataSonics Xpdr    -10.000  0.700 1.30    1 0.09100000  Dewey
#> 7  Benthos Xpdr        -5.000  0.000 1.00    1 0.07068583  Dewey
#> 8  Brancker TR1000     -0.448  0.270 1.30    1 0.01377000  Dewey
#> 9  Brancker XL-200     -0.430  0.540 1.30    1 0.03240000  Dewey
#> 10 Brancker XL-210      0.714  0.840 1.30    1 0.06384000  Dewey
#> 11 Heavy Sphere       -10.000  0.300 0.65    1 0.07068583  Dewey
#> 12 CDMS 10sqft Tow    -90.720  1.087 1.00    1 0.92800212  Dewey
#> 13 CDMS 30sqft Tow   -415.720  1.880 1.20    1 2.77591127  Dewey
#> 14 IceProfiler4+Flt    62.000  1.300 0.80    1 0.12566371  Dewey
#> 15 Heavy Sphere 2    -100.000  0.400 0.65    1 0.12566371  Dewey
#> 16 SBE 19+ cage       -20.000  0.450 1.30    1 0.09000000  Dewey
#> 17 McLane Mk8 SedTr   -20.000  1.200 1.30    1 0.36000000  Dewey
#> 18 BPS Package        -30.000  1.600 1.60    1 1.92000000  Dewey
#> 19 Oceanetic Pinger    -5.000  0.600 1.30    1 0.03000000  Dewey
#> 20 ONC-uCatADop       -10.000  0.900 1.30    1 0.13500000  Dewey
#> 21 McLane Mk78 SdTr   -25.000  1.640 1.30    1 0.67240000  Dewey
#> 22 SubOcean2d         -52.000  1.680 2.30    1 0.47040000  Dewey
#> 23 SubOcean2d2        -52.000  1.680 1.30    1 2.21670778  Dewey
#> 24 TowedCylinder      -10.000  1.000 1.30    1 0.20000000  Dewey
#> 25 TowedSphere        -10.000  0.500 0.65    1 0.19634954  Dewey
#>        originalName
#> 1   AanderaaT.chain
#> 2            SBE-16
#> 3     A. tide gauge
#> 4   Keil sedmt trap
#> 5         ULS#2/APL
#> 6   DataSonics Xpdr
#> 7      Benthos Xpdr
#> 8   Brancker TR1000
#> 9   Brancker XL-200
#> 10  Brancker XL-210
#> 11     Heavy Sphere
#> 12  CDMS 10sqft Tow
#> 13  CDMS 30sqft Tow
#> 14 IceProfiler4+Flt
#> 15   Heavy Sphere 2
#> 16     SBE 19+ cage
#> 17 McLane Mk8 SedTr
#> 18      BPS Package
#> 19 Oceanetic Pinger
#> 20     ONC-uCatADop
#> 21 McLane Mk78 SdTr
#> 22       SubOcean2d
#> 23      SubOcean2d2
#> 24    TowedCylinder
#> 25      TowedSphere
#> 
#> $releases
#>                          name buoyancy height  width   CD code source
#> 1                   EG&G 723a   -15.20  1.194 0.1173 0.65   NA    BIO
#> 2  EG&G 723a with tension bar   -25.87  1.320 0.1061 0.65   NA    BIO
#> 3                benthos 965a   -36.73  1.230 0.1033 0.65   NA    BIO
#> 4       benthos 966a and 866a   -13.16  0.680 0.1647 0.65   NA    BIO
#> 5                benthos 865a   -27.86  1.000 0.1990 0.65   NA    BIO
#> 6      vr2w with float collar   -12.11  0.400 0.5925 0.65   NA    BIO
#> 7               Oceano RT-121   -30.00  0.890 0.0000 1.30    1  Dewey
#> 8               Oceano RT-161   -35.00  0.890 0.0000 1.30    1  Dewey
#> 9                Oceano AR191   -26.00  0.850 0.0000 1.30    1  Dewey
#> 10               Oceano AR261   -21.00  0.820 0.0000 1.30    1  Dewey
#> 11                  EG&G 8242   -28.00  0.940 0.0000 1.30    1  Dewey
#> 12                double 8242   -68.00  0.950 0.0000 1.30    1  Dewey
#> 13              double RT-161   -70.00  0.890 0.0000 1.30    1  Dewey
#> 14              double AR-191   -60.00  0.850 0.0000 1.30    1  Dewey
#> 15             double AR/EG&G   -64.00  0.950 0.0000 1.30    1  Dewey
#> 16                    CDMS AR   -20.00  2.000 0.0000 1.20    1  Dewey
#> 17                Mesotech AR    -5.00  1.000 0.0000 1.30    7  Dewey
#> 18           InterOcean1090ED    -5.00  0.800 0.0000 1.30    1  Dewey
#> 19           Dble Mesotech501   -11.00  1.420 0.0000 1.20    1  Dewey
#> 20              Oceano AR2500   -22.00  0.830 0.0000 1.30    1  Dewey
#>                      originalName
#> 1                       EG&G 723A
#> 2      EG&G 723A WITH TENSION BAR
#> 3            BENTHOS 965A RELEASE
#> 4   BENTHOS 966A and 866A RELEASE
#> 5            BENTHOS 865A RELEASE
#> 6  VR2W RELEASE with float collar
#> 7                   Oceano RT-121
#> 8                   Oceano RT-161
#> 9                    Oceano AR191
#> 10                   Oceano AR261
#> 11                      EG&G 8242
#> 12                    double 8242
#> 13                  double RT-161
#> 14                  double AR-191
#> 15                 double AR/EG&G
#> 16                        CDMS AR
#> 17                    Mesotech AR
#> 18               InterOcean1090ED
#> 19               Dble Mesotech501
#> 20                  Oceano AR2500
#> 
#> $wires
#>                                      name buoyancyPerMeter  diameter  CD code
#> 1  3/16in galvanized wire coated to 1/4in          -0.0775 0.0063500 1.3   NA
#> 2  1/4in galvanized wire coated to 5/16in          -0.1380 0.0079375 1.3   NA
#> 3                            1/4in Kevlar          -0.0090 0.0070000 1.3    6
#> 4                           5/16in Kevlar          -0.0140 0.0080000 1.3    6
#> 5                            3/8in Kevlar          -0.0230 0.0090000 1.3    6
#> 6                           7/16in Kevlar          -0.0260 0.0110000 1.3    6
#> 7                            1/2in Kevlar          -0.0320 0.0130000 1.3    6
#> 8                           9/16in Kevlar          -0.0430 0.0140000 1.3    6
#> 9                            5/8in Kevlar          -0.0540 0.0160000 1.3    6
#> 10                       3/16in wire rope          -0.1200 0.0070000 1.3    1
#> 11                        1/4in wire rope          -0.1400 0.0070000 1.3    1
#> 12                       5/16in wire rope          -0.1900 0.0080000 1.3    1
#> 13                        3/8in wire rope          -0.3300 0.0090000 1.3    1
#> 14                        1/4in wire/jack          -0.1300 0.0080000 1.3    1
#> 15                       5/16in wire/jack          -0.1860 0.0110000 1.3    1
#> 16                        3/8in wire/jack          -0.3180 0.0130000 1.3    1
#> 17                        1/2in wire/jack          -0.5470 0.0170000 1.3    1
#> 18                             7/16in VLS          -0.0230 0.0110000 1.3    1
#> 19                              1/2in VLS          -0.0410 0.0130000 1.3    1
#> 20                            11mm Perlon          -0.0050 0.0110000 1.3    3
#> 21                          7/16in Dacron          -0.0320 0.0110000 1.3    3
#> 22                           1/2in Dacron          -0.0410 0.0130000 1.3    3
#> 23                          9/16in Dacron          -0.0580 0.0140000 1.3    3
#> 24                           3/4in Dacron          -0.0990 0.0190000 1.3    3
#> 25                            5/8in Nylon          -0.0220 0.0160000 1.3    2
#> 26                            3/4in Nylon          -0.0300 0.0190000 1.3    2
#> 27                                1 Nylon          -0.0500 0.0250000 1.3    2
#> 28                         3/4in Polyprop           0.0160 0.0190000 1.3    4
#> 29                     1/2in Neutral Wire           0.0000 0.0130000 1.5    6
#> 30                       1in Neutral Wire           0.0000 0.0250000 1.3    6
#> 31                     1/4in AmSteel-Blue           0.0020 0.0070000 1.3    3
#> 32                     3/8in AmSteel-Blue           0.0030 0.0090000 1.3    3
#> 33                     1/2in AmSteel-Blue           0.0060 0.0120000 1.3    3
#> 34                              BPS Cable          -0.1240 0.0070000 1.3    1
#> 35                     5/8in AmSteel-Blue           0.0090 0.0160000 1.3    3
#> 36                         BPS Power/Coms          -0.3840 0.0280000 1.3    6
#> 37                        1/4-1/2 WarpSpd          -0.0090 0.0090000 1.3    6
#> 38                            1/2 Dyneema           0.0200 0.0120000 1.3    8
#>    source                        originalName
#> 1     BIO 3/16" GALVANIZED WIRE COATED TO 1/4
#> 2     BIO 1/4" GALVANIZED WIRE COATED TO 5/16
#> 3   Dewey                          1/4 Kevlar
#> 4   Dewey                         5/16 Kevlar
#> 5   Dewey                          3/8 Kevlar
#> 6   Dewey                         7/16 Kevlar
#> 7   Dewey                          1/2 Kevlar
#> 8   Dewey                         9/16 Kevlar
#> 9   Dewey                          5/8 Kevlar
#> 10  Dewey                      3/16 wire rope
#> 11  Dewey                       1/4 wire rope
#> 12  Dewey                      5/16 wire rope
#> 13  Dewey                       3/8 wire rope
#> 14  Dewey                       1/4 wire/jack
#> 15  Dewey                      5/16 wire/jack
#> 16  Dewey                       3/8 wire/jack
#> 17  Dewey                       1/2 wire/jack
#> 18  Dewey                            7/16 VLS
#> 19  Dewey                             1/2 VLS
#> 20  Dewey                         11mm Perlon
#> 21  Dewey                         7/16 Dacron
#> 22  Dewey                          1/2 Dacron
#> 23  Dewey                         9/16 Dacron
#> 24  Dewey                          3/4 Dacron
#> 25  Dewey                           5/8 Nylon
#> 26  Dewey                           3/4 Nylon
#> 27  Dewey                             1 Nylon
#> 28  Dewey                        3/4 Polyprop
#> 29  Dewey                    1/2 Neutral Wire
#> 30  Dewey                    1in Neutral Wire
#> 31  Dewey                    1/4 AmSteel-Blue
#> 32  Dewey                    3/8 AmSteel-Blue
#> 33  Dewey                    1/2 AmSteel-Blue
#> 34  Dewey                           BPS Cable
#> 35  Dewey                    5/8 AmSteel-Blue
#> 36  Dewey                      BPS Power/Coms
#> 37  Dewey                     1/4-1/2 WarpSpd
#> 38  Dewey                         1/2 Dyneema
#>