Dataset of wire, chain, and float properties

This dataset was constructed from the mdcodes.mat (Dewey 1999, 2021). This is provided for use by wire(), chain() and float(). Users will seldom need to consult this table directly, but those that do should note that Dewey's lengths in cm have been converted to m.

data(mooringElements)

References

Dewey, Richard K. "Mooring Design & Dynamics-a Matlab® Package for Designing and Analyzing Oceanographic Moorings." Marine Models 1, no. 1 (December 1, 1999): 103–57. https://doi.org/10.1016/S1369-9350(00)00002-X Dewey, Richard. "Mooring Design and Dynamics." Accessed May 15, 2021. http://canuck.seos.uvic.ca/rkd/mooring/moordyn.php

Examples

library(mooring)
data(mooringElements)
mooringElements

#> $anchors
#>         name buoyancy height source
#> 1 trainwheel    -1000      0 kelley
#> 
#> $chains
#>                      name buoyancyPerMeter  width  CD code source
#> 1  5/8in galvanized chain           -4.323 0.0572 1.3   NA    BIO
#> 2          1/4in chain SL           -1.690 0.0150 1.3    1  Dewey
#> 3          3/8in chain SL           -2.330 0.0220 1.3    1  Dewey
#> 4          1/2in chain SL           -4.120 0.0300 1.3    1  Dewey
#> 5            1in chain SL          -13.000 0.0400 1.3    1  Dewey
#> 6          1/2in chain LL           -3.100 0.0300 1.3    1  Dewey
#> 7        3/8in buoy chain           -1.677 0.0200 1.3    1  Dewey
#> 8        1/2in buoy chain           -3.070 0.0250 1.3    1  Dewey
#> 9        5/8in buoy chain           -4.730 0.0350 1.3    1  Dewey
#> 10       3/4in buoy chain           -6.870 0.0400 1.3    1  Dewey
#> 11         1in buoy chain          -12.250 0.0500 1.3    1  Dewey
#> 
#> $connectors
#>                        name buoyancy height  width  CD code source
#> 1                    swivel   -0.953  0.000 0.0000 1.3   NA    BIO
#> 2            7/16in shackle   -0.150  0.000 0.0000 1.3   NA    BIO
#> 3    1/4in nicopress sleeve   -0.034  0.000 0.0000 1.3   NA    BIO
#> 4  5/16in heavy SS thimbles   -0.041  0.000 0.0000 1.3   NA    BIO
#> 5            ballast weight  -45.408  1.000 0.0254 1.3   NA    BIO
#> 6            shac-ring-shac   -0.750  0.300 0.0250 1.3    1  Dewey
#> 7           shac-3link-shac   -1.000  0.400 0.0200 1.3    1  Dewey
#> 8             1/2in SS link   -0.200  0.100 0.0250 1.3    1  Dewey
#> 9           1/2in galv link   -0.200  0.100 0.0250 1.3    1  Dewey
#> 10            1/2in shackle   -0.300  0.080 0.0250 1.3    1  Dewey
#> 11            3/8in shackle   -0.125  0.040 0.0250 1.3    1  Dewey
#> 12            5/8in shackle   -0.650  0.065 0.0500 1.3    1  Dewey
#> 13              1in shackle   -2.200  0.080 0.0300 1.3    1  Dewey
#> 14             1/2in swivel   -0.650  0.140 0.0500 1.3    1  Dewey
#> 15             3/4in swivel   -1.500  0.140 0.0400 1.3    1  Dewey
#> 16    1/2in shac+3/8in shac   -0.425  0.120 0.0250 1.3    1  Dewey
#> 17           3T SeineSwivel   -0.650  0.105 0.0400 1.3    1  Dewey
#> 18            3/4in shackle   -0.880  0.060 0.0250 1.3    1  Dewey
#> 19          shac-Pring-shac   -0.500  0.200 0.0200 1.3    1  Dewey
#> 
#> $floats
#>                                    name buoyancy height diameter   CD code
#> 1        streamlined bub 2 x 17in glass    45.41 0.7400   0.4550 0.60   NA
#> 2      a2 package adcp and 2 Viny balls    29.08 0.5630   0.3970 0.60   NA
#> 3              3 pack Viny 12b-3 floats    57.14 1.0000   0.5290 1.10   NA
#> 4          streamlined bub 3 Viny balls    52.65 0.5630   0.3970 0.60   NA
#> 5                      bub 2x17in glass    43.78 1.1870   0.8390 1.10   NA
#> 6             IPS / 2x b3 subs assembly    78.88 1.0000   0.7480 0.60   NA
#> 7            ADCP / 2x c3 subs assembly    42.24 1.0000   0.9100 0.60   NA
#> 8   stablemoor 533lb 3500 msw with ADCP   192.45 1.2860   1.7270 0.60   NA
#> 9  stablemoor 1000lb 3500 msw with ADCP   404.39 1.2860   1.7270 0.60   NA
#> 10 stablemoor 1015lb 1500 msw with ADCP   411.22 1.2860   1.7200 0.60   NA
#> 11  stablemoor 580lb 1500 msw with ADCP   213.78 1.2860   1.7270 0.60   NA
#> 12    syntactic float with ADCP bracket   839.18 2.4200   1.2540 0.60   NA
#> 13    syn. float,bracket and 109lb ADCP   789.80 2.4200   1.2540 0.60   NA
#> 14                             61in ORE   999.00 1.6500   1.6500 0.65    1
#> 15                             48in ORE   620.00 1.2340   1.2340 0.65    1
#> 16                             41in ORE   433.00 1.0400   1.0400 0.65    1
#> 17                             37in ORE   300.00 0.9400   0.9400 0.65    1
#> 18                             28in ORE   133.00 0.7100   0.7100 0.65    1
#> 19                           30in float   160.00 0.7600   0.7600 0.65    1
#> 20                         Kiel SFS40in   320.00 1.0000   1.0000 0.65    1
#> 21                      Torp. flt, 3-28   364.00 0.8300   0.8300 0.75    1
#> 22                           17in glass    23.00 0.5700   0.5700 0.95    1
#> 23                        Billings 12in    10.00 0.3430   0.3430 0.95    1
#> 24                         BENTHOS 17in    23.00 0.5100   0.5100 0.95    1
#> 25                            16in Viny    19.00 0.4000   0.4000 0.60    1
#> 26                          Double 17in    48.00 0.5000   0.6500 0.95    1
#> 27                     Double 16in Viny    38.00 0.5000   0.6000 0.60    1
#> 28                       Trpl 16in Viny    56.00 1.5000   0.7500 0.60    1
#> 29                      Trpl 12in glass    30.00 0.9500   0.6200 1.00    1
#> 30                     CDMS 1.2m sphere   682.00 1.2000   1.2000 0.50    1
#> 31                       CDMS 2m sphere  2470.00 2.0000   2.0000 0.50    1
#> 32                       54cm Alum CAP3    61.00 0.5500   0.5400 0.65    1
#> 33                     54x30inClamShell   405.00 1.4000   0.7500 0.15    1
#> 34                      Top of Tow Rope     0.01 0.0000   0.0000 0.00    1
#> 35                            BPS Float  8000.00 1.5000   0.0000 1.60    1
#> 36                       ADCP FloatTech   460.00 1.4200   1.4200 0.65    1
#> 37                      FloatTech CF-12    11.00 0.3000   0.0000 1.00    1
#> 38                     Tripple FT CF-12    33.00 0.9000   0.0000 1.00    1
#> 39                    Geodyn 8ft Toroid  1950.00 0.7620   0.0000 0.65    3
#> 40                               HMB 20    34.50 0.5080   0.5080 0.65   NA
#> 41                               HMB 25    68.50 0.6350   0.6350 0.65   NA
#> 42                               HMB 31   141.00 0.7874   0.7874 0.65   NA
#> 43                               HMB 36   225.00 0.9144   0.9144 0.65   NA
#> 44                               HMB 40   300.00 1.0160   1.0160 0.65   NA
#> 45                               HMB 44   413.00 1.1176   1.1176 0.65   NA
#> 46                               HMB 49   565.00 1.2446   1.2446 0.65   NA
#> 47                               HMB 51   633.00 1.2954   1.2954 0.65   NA
#> 48                               HMB 56   831.00 1.4224   1.4224 0.65   NA
#> 49                               HMB 62  1122.00 1.5748   1.5748 0.65   NA
#>     source
#> 1      BIO
#> 2      BIO
#> 3      BIO
#> 4      BIO
#> 5      BIO
#> 6      BIO
#> 7      BIO
#> 8      BIO
#> 9      BIO
#> 10     BIO
#> 11     BIO
#> 12     BIO
#> 13     BIO
#> 14   Dewey
#> 15   Dewey
#> 16   Dewey
#> 17   Dewey
#> 18   Dewey
#> 19   Dewey
#> 20   Dewey
#> 21   Dewey
#> 22   Dewey
#> 23   Dewey
#> 24   Dewey
#> 25   Dewey
#> 26   Dewey
#> 27   Dewey
#> 28   Dewey
#> 29   Dewey
#> 30   Dewey
#> 31   Dewey
#> 32   Dewey
#> 33   Dewey
#> 34   Dewey
#> 35   Dewey
#> 36   Dewey
#> 37   Dewey
#> 38   Dewey
#> 39   Dewey
#> 40  Kelley
#> 41  Kelley
#> 42  Kelley
#> 43  Kelley
#> 44  Kelley
#> 45  Kelley
#> 46  Kelley
#> 47  Kelley
#> 48  Kelley
#> 49  Kelley
#> 
#> $instruments
#>                                                 name buoyancy height  area   CD
#> 1                    seacat 16-03 (bar,plastic case)    -7.49 0.8400 0.122 0.65
#> 2              seacat 16-04 (bar,titanium no press.)   -10.89 0.8763 0.062 0.65
#> 3            seacat 16-04 (bar,titanium with press.)   -13.39 1.0668 0.077 0.65
#> 4                  seabird CTD (ios oxygen with bar)   -18.16 1.2700 0.094 0.65
#> 5                                  DVS current meter    -4.54 0.7000 0.243 0.65
#> 6                                        wotan (bar)   -20.42 1.6800 0.176 0.65
#> 7                                       RDI with bar   -45.39 2.1800 0.270 0.65
#> 8                                     RCM-8 with fin   -21.33 0.7400 0.060 0.65
#> 9                                    RCM-11 in frame   -20.31 0.8000 0.078 0.65
#> 10                          seaguard (2000 & 6000 m)   -19.59 0.8600 0.083 0.65
#> 11                                 vemco on a ss bar    -2.22 1.2000 0.052 0.65
#> 12                        SBE37 microcat on a ss bar    -4.65 1.2000 0.060 0.65
#> 13                             sediment trap (#1349)   -61.27 1.8290 1.061 0.65
#> 14                     SBE37 microcat clamp-on style    -1.53 0.5590 0.117 0.65
#> 15  WH ADCP sentinel (orange) 500m with inline frame   -14.69 0.7600 0.091 0.65
#> 16 WH ADCP sentinel (yellow) 6000m with inline frame   -22.04 0.7900 0.111 0.65
#>    code source
#> 1    NA    BIO
#> 2    NA    BIO
#> 3    NA    BIO
#> 4    NA    BIO
#> 5    NA    BIO
#> 6    NA    BIO
#> 7    NA    BIO
#> 8    NA    BIO
#> 9    NA    BIO
#> 10   NA    BIO
#> 11   NA    BIO
#> 12   NA    BIO
#> 13   NA    BIO
#> 14   NA    BIO
#> 15   NA    BIO
#> 16   NA    BIO
#> 
#> $releases
#>                             name buoyancy height  width   CD code source
#> 1                      EG&G 723a   -15.20  1.194 0.1173 0.65   NA    BIO
#> 2     EG&G 723a with tension bar   -25.87  1.320 0.1061 0.65   NA    BIO
#> 3           benthos 965a release   -36.73  1.230 0.1033 0.65   NA    BIO
#> 4  benthos 966a and 866a release   -13.16  0.680 0.1647 0.65   NA    BIO
#> 5           benthos 865a release   -27.86  1.000 0.1990 0.65   NA    BIO
#> 6 vr2w release with float collar   -12.11  0.400 0.5925 0.65   NA    BIO
#> 
#> $wires
#>                                      name buoyancyPerMeter  diameter  CD code
#> 1  3/16in galvanized wire coated to 1/4in          -0.0775 0.0063500 1.3   NA
#> 2  1/4in galvanized wire coated to 5/16in          -0.1380 0.0079375 1.3   NA
#> 3                            1/4in Kevlar          -0.0090 0.0070000 1.3    6
#> 4                           5/16in Kevlar          -0.0140 0.0080000 1.3    6
#> 5                            3/8in Kevlar          -0.0230 0.0090000 1.3    6
#> 6                           7/16in Kevlar          -0.0260 0.0110000 1.3    6
#> 7                            1/2in Kevlar          -0.0320 0.0130000 1.3    6
#> 8                           9/16in Kevlar          -0.0430 0.0140000 1.3    6
#> 9                            5/8in Kevlar          -0.0540 0.0160000 1.3    6
#> 10                       3/16in wire rope          -0.1200 0.0070000 1.3    1
#> 11                        1/4in wire rope          -0.1400 0.0070000 1.3    1
#> 12                       5/16in wire rope          -0.1900 0.0080000 1.3    1
#> 13                        3/8in wire rope          -0.3300 0.0090000 1.3    1
#> 14                        1/4in wire/jack          -0.1300 0.0080000 1.3    1
#> 15                       5/16in wire/jack          -0.1860 0.0110000 1.3    1
#> 16                        3/8in wire/jack          -0.3180 0.0130000 1.3    1
#> 17                        1/2in wire/jack          -0.5470 0.0170000 1.3    1
#> 18                             7/16in VLS          -0.0230 0.0110000 1.3    1
#> 19                              1/2in VLS          -0.0410 0.0130000 1.3    1
#> 20                            11mm Perlon          -0.0050 0.0110000 1.3    3
#> 21                          7/16in Dacron          -0.0320 0.0110000 1.3    3
#> 22                           1/2in Dacron          -0.0410 0.0130000 1.3    3
#> 23                          9/16in Dacron          -0.0580 0.0140000 1.3    3
#> 24                           3/4in Dacron          -0.0990 0.0190000 1.3    3
#> 25                            5/8in Nylon          -0.0220 0.0160000 1.3    2
#> 26                            3/4in Nylon          -0.0300 0.0190000 1.3    2
#> 27                              1in Nylon          -0.0500 0.0250000 1.3    2
#> 28                         3/4in Polyprop           0.0160 0.0190000 1.3    4
#> 29                     1/2in Neutral Wire           0.0000 0.0130000 1.5    6
#> 30                       1in Neutral Wire           0.0000 0.0250000 0.0    6
#> 31                     1/4in AmSteel-Blue           0.0020 0.0070000 1.3    3
#> 32                     3/8in AmSteel-Blue           0.0030 0.0090000 1.3    3
#> 33                     1/2in AmSteel-Blue           0.0060 0.0120000 1.3    3
#> 34                              BPS Cable          -0.1240 0.0070000 1.3    1
#> 35                     5/8in AmSteel-Blue           0.0090 0.0160000 1.3    3
#> 36                         BPS Power/Coms          -0.3840 0.0280000 1.3    6
#> 37                        1/4-1/2 WarpSpd          -0.0090 0.0090000 1.3    6
#> 38                            1/2 Dyneema           0.0200 0.0120000 1.3    8
#>    source
#> 1     BIO
#> 2     BIO
#> 3   Dewey
#> 4   Dewey
#> 5   Dewey
#> 6   Dewey
#> 7   Dewey
#> 8   Dewey
#> 9   Dewey
#> 10  Dewey
#> 11  Dewey
#> 12  Dewey
#> 13  Dewey
#> 14  Dewey
#> 15  Dewey
#> 16  Dewey
#> 17  Dewey
#> 18  Dewey
#> 19  Dewey
#> 20  Dewey
#> 21  Dewey
#> 22  Dewey
#> 23  Dewey
#> 24  Dewey
#> 25  Dewey
#> 26  Dewey
#> 27  Dewey
#> 28  Dewey
#> 29  Dewey
#> 30  Dewey
#> 31  Dewey
#> 32  Dewey
#> 33  Dewey
#> 34  Dewey
#> 35  Dewey
#> 36  Dewey
#> 37  Dewey
#> 38  Dewey
#>